lee
20210219

+++ controllers/videoController.js
@@ -1,6 +1,7 @@
 
 import routes from "../routes";
-import Video from "../models/Video"
+import Video from "../models/Video";
+import Comment from "../models/Comment";
 //render라는 함수가 views 폴더에서pug형식에 home이라는 이름의 파일을 찾앚준다
 
 export const home= async(req,res)=>{
@@ -37,8 +38,11 @@ export const postUpload=async(req,res)=>{
     const newVideo = await Video.create({
         fileUrl:path,
         title,
-        description
+        description,
+        creator:req.user.id
     });
+    req.user.videos.push(newVideo.id);
+    req.user.save();
         // To Do: Upload and save video
     res.redirect(routes.videoDetail(newVideo.id));
 };
@@ -48,7 +52,7 @@ export const videoDetail=async(req,res)=>{
         params:{id}
     }=req;
     try{
-        const video= await Video.findById(id);
+        const video= await Video.findById(id).populate("creator").populate("comments");
         res.render("videoDetail", {pageTitle:video.title, video});
     }catch(error){
         console.log(error);
@@ -62,7 +66,11 @@ export const getEditVideo=async(req,res)=>{
     }=req;
     try{
         const video= await Video.findById(id);
-        res.render("editVideo", {pageTitle:`Edit ${video.title}`, video});
+        if(video.creator.toString()!==req.user.id){
+            throw Error();
+        }else{
+            res.render("editVideo", {pageTitle:`Edit ${video.title}`, video});
+        }
     }catch(error){
         res.redirect(routes.home);
     }
@@ -84,9 +92,56 @@ export const deleteVideo=async(req,res)=>{
         params:{id}
     }=req;
     try{
-        await Video.findOneAndRemove({_id:id});
+        const video= await Video.findById(id);
+        if(video.creator.toString()!==req.user.id){
+            throw Error();
+        }else{
+            await Video.findOneAndRemove({_id:id});
+        }
     }catch(error){
         console.log(error);
     }
     res.redirect(routes.home);
 }
+
+
+export const postRegisterView=async(req,res)=>{
+    const{
+        params:{id}
+    }=req;
+    try{
+        const video=await Video.findById(id);
+        video.views+=1;
+        video.save();
+        res.status(200);
+    }catch(error){
+        res.status(400);
+        
+    }finally{
+        res.end();
+    }
+};
+
+//Add Comment
+
+export const postAddComment= async(req,res)=>{
+    const{
+        params:{id},
+        body:{comment},
+        user
+    }=req;
+    try{
+        const video=await Video.findById(id);
+        const newComment =await Comment.create({
+            text:comment,
+            creator:user.id
+        });
+        video.comments.push(newComment.id);
+        video.save();
+    }catch(error){
+        res.status(400);
+
+    }finally{
+        res.end();
+    }
+};
